Use aktivitetskravfilter for test veileder
AND test the same way for active aktivitetskrav at the application level
as in the database level.

This should get rid of the Berit bug.
